cambios

+++ Laragon/www/lelsystem/LELSYSTEM-V2/database/migrations/2022_08_24_182099_create_cursos_table.php
@@ -20,7 +20,7 @@ class CreateCursosTable extends Migration
             $table->date('fecha_inicio');
             $table->date('fecha_fin');
             $table->boolean('activo');
-            $table->bigInteger('id_docente');
+            $table->integer('id_docente');
             $table->foreign('id_docente')->references('id_docente')->on('docentes');
         });
     }
